fix some test
